chore(cli): Refactor
