+++ src/cli/options.ts
@@ -1,38 +1,6 @@
+import { ToolConfig, OutputFormat } from "./types";
 import { Severity, parseSeverity } from "../internals/warnings";
 import { Option } from "commander";
-
-export const DUMP_STDOUT_PATH = "-";
-
-/**
- * Configuration for a tool.
- */
-export interface ToolConfig {
-  /**
-   * Name of the class that implements the tool.
-   */
-  className: string;
-  /**
-   * Generic set of options for tools
-   */
-  options?: Record<string, unknown>;
-}
-
-/**
- * Configuration for a detector.
- */
-export interface DetectorConfig {
-  /**
-   * Path to the module containing the detector.
-   * Used only for custom out-of-tree detectors.
-   */
-  modulePath?: string;
-  /**
-   * Name of the class that implements the detector.
-   */
-  className: string;
-}
-
-export type OutputFormat = "json" | "plain";
 
 export interface CLIOptions {
   tools?: ToolConfig[];
